backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
PyJWT==2.9.0
pymongo==4.10.1
Werkzeug==3.0.3
python-dotenv==1.0.1

